+++ spinetoolbox/spine_db_editor/widgets/custom_qwidgets.py
@@ -85,19 +85,14 @@ class OpenFileButton(QWidget):
             f"""
             QToolButton {{
                 padding-left: 16px; padding-right: {16 + menu_button_size}px; padding-top: 6px; padding-bottom: 6px;
-                background-color: #fff;
-                border: 1px solid #ccc;
+                border: 1px solid;
                 border-style: outset;
             }}
-            QToolButton:hover {{
-                background-color: #eee;
-            }}
             QToolButton:pressed {{
-                background-color: #ddd;
                 border-style: inset;
             }}
             QToolButton::menu-button {{
-                border: 1px solid #ccc;
+                border: 1px solid;
                 border-style: outset;
                 width: {menu_button_size}px;
             }}
@@ -106,8 +101,7 @@ class OpenFileButton(QWidget):
             }}
             QProgressBar {{
                 text-align: center;
-                background-color: #fff;
-                border: 1px solid #ccc;
+                border: 1px solid;
                 border-style: inset;
             }}
             """