+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 
 ### Added
 
+- Added light and dark themes. Detected from OS settings.
+
 ### Changed
 
 ### Deprecated
@@ -48,6 +50,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
   in addition to scenario and alternative descriptions.
   Note, that the feature is not available for existing mappings.
   You need to create new *Entity class* or type mappings with *Parameter type* set to None or Default value.
+
+### Changed
+
+### Deprecated
 
 ### Removed